+++ src/duplocloud/client.py
@@ -58,7 +58,8 @@ class DuploClient():
                query: args.QUERY=None,
                output: args.OUTPUT="json",
                loglevel: args.LOGLEVEL="WARN",
-               wait: args.WAIT=False):
+               wait: args.WAIT=False,
+               wait_timeout: args.WAIT_TIMEOUT=None):
     """DuploClient Constructor
     
     Creates an instance of a duplocloud client configured for a certain portal. All of the arguments are optional and can be set in the environment or in the config file. The types of each of the arguments are annotated types that are used by argparse to create the command line arguments.
@@ -117,6 +118,7 @@ class DuploClient():
     self.loglevel = loglevel
     self.logger = self.logger_for()
     self.wait = wait
+    self.wait_timeout = wait_timeout
 
   @staticmethod
   def from_env():